add323

+++ src/main/java/xyz/coolestme/community/service/QuestionService.java
@@ -64,7 +64,7 @@ public class QuestionService {
         return pageinationDTO;
     }
 
-    public PageinationDTO list(Integer userId, Integer page, Integer size) {
+    public PageinationDTO list(Long userId, Integer page, Integer size) {
         PageinationDTO pageinationDTO = new PageinationDTO();
 
         QuestionExample questionExample = new QuestionExample();
@@ -102,7 +102,7 @@ public class QuestionService {
         return pageinationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question =questionMapper.selectByPrimaryKey(id);
         if (question == null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -137,7 +137,7 @@ public class QuestionService {
         }
     }
 
-    public void incView(Integer id) {
+    public void incView(Long id) {
         Question question = new Question();
         question.setId(id);
         question.setViewCount(1);
